Refactor file structure for readable code
