+++ benches/read_write.rs
@@ -1,7 +1,7 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 use serde_json::json;
 use smoldb::storage::{
-    content_manager::{Collection, CollectionConfig},
+    collection::{Collection, CollectionConfig},
     segment::{Point, PointId},
 };
 use tempfile::TempDir;
