+++ src/lib.rs
@@ -3,3 +3,4 @@ pub mod args;
 pub mod channel_service;
 pub mod consensus;
 pub mod storage;
+pub mod types;
