+++ src/storage/mod.rs
@@ -1,6 +1,5 @@
-pub mod content_manager;
+pub mod collection;
 pub mod error;
 pub mod replicas;
 pub mod segment;
-pub mod shard;
-pub mod shard_trait;
+pub mod toc;
