+++ src/types.rs
@@ -1,0 +1,3 @@
+pub type SegmentId = u64;
+pub type ShardId = u32;
+pub type PeerId = u64;
